refactor: reorganize src into a single consistent structure
The lib/ folder had accumulated three competing organizing schemes over
multiple iterations (by-feature, by-type, and loose top-level files), so
there was no single rule for where a file belonged. Blog code was split
across two trees, feature components lived in three different places, and
folders like shared/ and index/ were misleadingly named.

Adopt one rule: $lib holds only shared/cross-page code; a component used by
exactly one page lives next to that page's route.

- Colocate single-page components with their routes (home, about, career,
  blog list/pagination)
- Consolidate mdsvex layouts into lib/layouts/ (Post/Now/Colophon)
- Promote the general-purpose Tag into lib/components/; move Header/Footer
  there too (no longer loose)
- Rename lib/shared/ -> lib/stores/; flatten lib/utils/utils.js -> lib/utils.js
- Delete unreferenced legacy static/images/* assets and a stale
  mainImageThumb frontmatter line from an old iteration
- Update CLAUDE.md and mdsvex.config.js to match

Verified: npm run check (0/0), build, and 21/21 Playwright tests all pass.
