+++ src/blog/functional-programming-javascript.md
@@ -7,7 +7,6 @@ tags: [javascript, paradigms]
 mainImage: https://res.cloudinary.com/antonio-rossi/image/upload/w_1000,fl_progressive/v1648646194/articles/functional-programming/pure_zyijgf.jpg
 mainImageWebP: https://res.cloudinary.com/antonio-rossi/image/upload/w_1000,fl_progressive,f_webp/v1648646194/articles/functional-programming/pure_zyijgf.jpg
 mainImageAlt: An image of pure water with wite sand
-mainImageThumb: $lib/assets/article-thumbs/pure.jpg
 published: false
 date: 2022-03-30
 layout: development
