+++ src/routes/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-  import Hero from "$lib/index/Hero.svelte";
-  import Contacts from "$lib/index/Contacts.svelte";
-  import Articles from "$lib/index/Articles/Articles.svelte";
+  import Hero from "./Hero.svelte";
+  import Contacts from "./Contacts.svelte";
+  import Articles from "./Articles.svelte";
   let { data } = $props();
 </script>
