+++ src/routes/Hero.svelte
