+++ src/routes/about/+page.svelte
@@ -1,8 +1,8 @@
 <script>
-  import Skill from "$lib/components/about/Skill.svelte";
+  import Skill from "./Skill.svelte";
   import meImage from "$lib/assets/images/home/me-b-and-w.jpg?enhanced&w=300&h=300&quality=50";
-  import { calculateExperience } from "$lib/utils/utils.js";
-  import { calculatePastExperience } from "$lib/utils/utils.js";
+  import { calculateExperience } from "$lib/utils.js";
+  import { calculatePastExperience } from "$lib/utils.js";
 
   const skills = [
     { name: "React", start: "2020-01-01" },
